HOTEL-8: add currency exchange logic

+++ hotel-booking-system/src/main/java/org/example/model/Booking.java
@@ -15,7 +15,7 @@ public class Booking {
     private int nights;
     private MonetaryAmount totalPrice;
 
-    int calculateTotalPrice() {
+    double calculateTotalPrice() {
         return room.getPricePerNight().getAmount() * nights;
     }
 }
